Move AppText component to elements dir
